Fluent mapper factory API
